Doing some cleaning up
